requirements.txt
# ============================================================
# Multi-Agent Research & Reporting Platform — Dependencies
# ============================================================

# Orchestration
agent-framework>=1.0.0b0

# API
fastapi>=0.111.0
uvicorn>=0.29.0

# Azure Queue Storage
azure-storage-queue>=12.9.0

# Azure Blob Storage
azure-storage-blob>=12.19.0

# Azure SQL (async ODBC)
aioodbc>=0.5.0
pyodbc>=5.1.0

# Azure AI Foundry
azure-ai-projects>=1.0.0b1

# Azure AI Search
azure-search-documents>=11.6.0

# Azure Identity
azure-identity>=1.15.0

# Monitoring & Telemetry
azure-monitor-opentelemetry-exporter>=1.0.0b21
opentelemetry-sdk>=1.24.0

# Schema enforcement
pydantic>=2.7.0
pydantic-settings>=2.2.0

# HTTP client
httpx>=0.27.0

# Retry
tenacity>=8.3.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
